+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added `NFData` and `Generic` instances for `Comp`
 * Added `Prod`
 * `Data.Extensible.Class` now exports `compareMembership`
+* Renamed `runMembership` to `leadership`
 
 0.4.2
 -------------------------------------------------